Added footer button functionality
